requirements.txt
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
Django==3.1.3
django-crispy-forms==1.10.0
django-debug-toolbar==2.2
django-progressive-web-app==0.1.1
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
Pillow==8.0.1
protobuf==3.14.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
